Project-7/backend/requirements.txt
flask==2.3.0
langchain==0.1.0
google-generativeai==0.3.0
python-dotenv==1.0.0
flask-cors==4.0.0
chromadb==0.4.0
openai==1.3.0
PyPDF2==3.0.0
pdfplumber==0.10.0
sentence-transformers==2.2.0
